style: format files and move pages folder
Formatted all files and moved the pages folder to inside the src folder
