+++ src/components/Container/index.tsx
@@ -1,6 +1,6 @@
-import type { FC } from "react";
+import type { FC } from 'react';
 
-import styles from "./Container.module.scss";
+import styles from './Container.module.scss';
 
 const Container: FC = function ContainerElement({ children }) {
   return <div className={styles.container}>{children}</div>;
