+++ src/components/Footer/index.tsx
@@ -1,6 +1,6 @@
-import styles from "./Footer.module.scss";
+import styles from './Footer.module.scss';
 
-import Container from "../Container";
+import Container from '../Container';
 
 function Footer() {
   return (
