+++ src/pages/404.tsx
@@ -1,4 +1,4 @@
-import type { NextPage } from "next";
+import type { NextPage } from 'next';
 
 const Error404: NextPage = () => {
   return <div></div>;
