+++ src/pages/works/cosmo-mapeamento.tsx
@@ -1,10 +1,10 @@
-import type { NextPage } from "next";
-import styles from "../../src/styles/pages/Works.module.scss";
+import type { NextPage } from 'next';
+import styles from '../../styles/pages/Works.module.scss';
 
-import Head from "next/head";
-import Container from "../../src/components/Container";
-import CardList from "../../src/components/CardList";
-import Footer from "../../src/components/Footer";
+import Head from 'next/head';
+import Container from '../../components/Container';
+import CardList from '../../components/CardList';
+import Footer from '../../components/Footer';
 
 const CosmoMapeamento: NextPage = () => {
   return (
