+++ src/styles/global/index.scss
@@ -1,4 +1,4 @@
-@use "./reset.css";
-@use "./variables.css";
-@use "./elements.scss";
-@use "./utilities.scss";
+@use './reset.css';
+@use './variables.css';
+@use './elements.scss';
+@use './utilities.scss';
