+++ src/styles/global/utilities.scss
@@ -16,7 +16,7 @@
   position: relative;
 
   &::after {
-    content: "";
+    content: '';
     position: absolute;
     top: 100%;
     left: 0;
