+++ src/styles/pages/Home.module.scss
@@ -1,4 +1,4 @@
-@use "../partials/screen-breakpoints" as sb;
+@use '../partials/screen-breakpoints' as sb;
 
 // Header
 .header {
@@ -108,7 +108,7 @@
   position: relative;
 
   &::before {
-    content: "";
+    content: '';
     background-image: url(/svg/forward-icon.svg);
     background-repeat: no-repeat;
     background-position: center right;
@@ -126,7 +126,7 @@
   margin-left: var(--icon-position);
 
   &::before {
-    content: "";
+    content: '';
     background-image: url(/svg/email-icon.svg);
     background-repeat: no-repeat;
     background-position: center left;
